updated send parcel page and added new utility functions and others
