+++ src/lib/getDistrictData.js
@@ -1,0 +1,6 @@
+import { dbConnect } from "./dbConnect";
+
+export async function districtsData() {
+    const allDistricts = await dbConnect("districts").find().toArray();
+    return allDistricts;
+}
